+++ packages/vaster_runtime/lib/vaster_runtime.dart
@@ -4,6 +4,7 @@ library;
 export 'src/call_stack.dart';
 export 'src/cache_hint_tracker.dart';
 export 'src/decision_arbiter.dart';
+export 'src/extract_outcome.dart';
 export 'src/hitl_controller.dart';
 export 'src/policy_guard.dart';
 export 'src/register_file.dart';